feat(ui): rename feature flag reporting tab to insights
