+++ modules/front-end/src/app/features/safe/feature-flags/details/details.component.html
@@ -22,9 +22,9 @@
       </ng-template>
     </nz-tab>
     <nz-tab>
-      <a *nzTabLink nz-tab-link [routerLink]="['./reporting']" i18n="@@ff.routing.details.reporting">Reporting</a>
+      <a *nzTabLink nz-tab-link [routerLink]="['./insights']" i18n="@@ff.routing.details.insights">Insights</a>
       <ng-template nz-tab>
-        <ff-reporting></ff-reporting>
+        <insights></insights>
       </ng-template>
     </nz-tab>
   </nz-tabset>
