+++ modules/front-end/src/app/features/safe/feature-flags/details/insights/insights.component.html
@@ -11,7 +11,7 @@
   <div class="content-container">
     <div class="standard-div" name="detail">
       <div class="header">
-        <div class="title" i18n="@@ff.reporting.times-of-call">Times of call</div>
+        <div class="title" i18n="@@ff.reporting.flag-evaluations">Flag evaluations</div>
       </div>
       <div style="text-align: center;">{{usage}}</div>
       <div class="chart-container" style="width:97%">
