+++ modules/front-end/src/app/features/safe/feature-flags/details/insights/insights.component.less
@@ -32,12 +32,11 @@
         margin-bottom: 16px;
 
         .title {
-          font-family: Noto Sans S Chinese medium;
-          font-style: normal;
-          font-weight: 500;
+          font-weight: 600;
           font-size: 16px;
           line-height: 150%;
           color: @grey70-color;
+          height: 30px;
         }
       }
