+++ modules/front-end/src/app/features/safe/feature-flags/details/insights/insights.component.ts
@@ -6,21 +6,20 @@ import {IVariation} from "@shared/rules";
 import {EnvUserService} from "@services/env-user.service";
 import {Subject} from "rxjs";
 import {debounceTime} from "rxjs/operators";
-import {getPathPrefix, uuidv4} from "@utils/index";
+import {uuidv4} from "@utils/index";
 import {
-  IFeatureFlagEndUser,
   IFeatureFlagEndUserPagedResult,
   IntervalType,
   PeriodOption,
-  ReportFilter
-} from "@features/safe/feature-flags/details/reporting/types";
+  InsightsFilter
+} from "@features/safe/feature-flags/details/insights/types";
 
 @Component({
-  selector: 'ff-reporting',
-  templateUrl: './reporting.component.html',
-  styleUrls: ['./reporting.component.less']
+  selector: 'insights',
+  templateUrl: './insights.component.html',
+  styleUrls: ['./insights.component.less']
 })
-export class ReportingComponent implements OnInit {
+export class InsightsComponent implements OnInit {
 
   public usage: string = '';
 
@@ -56,7 +55,7 @@ export class ReportingComponent implements OnInit {
     });
 
     this.route.paramMap.subscribe(paramMap => {
-      this.filter = new ReportFilter(decodeURIComponent(paramMap.get('key')));
+      this.filter = new InsightsFilter(decodeURIComponent(paramMap.get('key')));
       this.featureFlagService.getByKey(this.filter.featureFlagKey).subscribe((res) => {
         this.variations = [{ id: this.featureFlagVariationAllId, value: $localize `:@@common.all:All`}, ...res.variations];
         this.filter.variationId = this.featureFlagVariationAllId;
@@ -66,7 +65,7 @@ export class ReportingComponent implements OnInit {
     });
   }
 
-  filter: ReportFilter;
+  filter: InsightsFilter;
 
   intervalTypes: any[];
   private setIntervalTypes() {
